requirements.txt
Django
Pillow
django-on-heroku
gunicorn

python-dotenv
python-decouple

#aws S3
boto3
django-storages

#multilang
django-modeltranslation
